+++ ui/apps/platform/src/services/GroupsService.ts
@@ -1,13 +1,21 @@
+import { Group } from 'types/group.proto';
 import axios from './instance';
-import { GetDefaultGroupRequest, Group, GroupBatchUpdateRequest } from '../types/group.proto';
 
 const url = '/v1/groups';
 const updateUrl = '/v1/groupsbatch';
 
+export type GetDefaultGroupRequest = {
+    authProviderId: string;
+    roleName: string;
+};
+
+export type GroupBatchUpdateRequest = {
+    previousGroups: Group[];
+    requiredGroups: Group[];
+};
+
 /**
  * Fetches list of groups of rules
- *
- * @returns {Promise<Object, Error>} fulfilled with array of groups
  */
 export function fetchGroups(): Promise<{ response: { groups: Group[] } }> {
     return axios.get(url).then((response) => ({
@@ -17,8 +25,6 @@ export function fetchGroups(): Promise<{ response: { groups: Group[] } }> {
 
 /**
  * Update/Add a group rule.
- *
- * @returns {Promise<Object, Error>}
  */
 export function updateOrAddGroup(request: GroupBatchUpdateRequest) {
     return axios.post(updateUrl, request);
@@ -26,8 +32,6 @@ export function updateOrAddGroup(request: GroupBatchUpdateRequest) {
 
 /**
  * Fetches the default rule.
- *
- * @returns {Promise<Object, Error>} fulfilled default group
  */
 export function getDefaultGroup({ authProviderId, roleName }: GetDefaultGroupRequest) {
     // The default group is characterized by the following:
@@ -43,8 +47,6 @@ export function getDefaultGroup({ authProviderId, roleName }: GetDefaultGroupReq
 
 /**
  * Deletes a group rule.
- *
- * @returns {Promise<Object, Error>}
  */
 export function deleteRuleGroup(data: Group) {
     const { key, authProviderId, value, id } = data.props;
